refactor: 궁성 내 이동 규칙을 가진 PalaceRestrictMoveStrategy로 통합
